[FLINK-4660] Allow ProcessFunction on DataStream
Introduce new ProcessOperator for this. Rename the pre-existing
ProcessOperator to KeyedProcessOperator.
