+++ flink-streaming-java/src/main/java/org/apache/flink/streaming/api/datastream/KeyedStream.java
@@ -40,7 +40,7 @@ import org.apache.flink.streaming.api.functions.query.QueryableValueStateOperato
 import org.apache.flink.streaming.api.functions.sink.SinkFunction;
 import org.apache.flink.streaming.api.graph.StreamGraphGenerator;
 import org.apache.flink.streaming.api.operators.OneInputStreamOperator;
-import org.apache.flink.streaming.api.operators.ProcessOperator;
+import org.apache.flink.streaming.api.operators.KeyedProcessOperator;
 import org.apache.flink.streaming.api.operators.StreamGroupedFold;
 import org.apache.flink.streaming.api.operators.StreamGroupedReduce;
 import org.apache.flink.streaming.api.transformations.OneInputTransformation;
@@ -187,6 +187,7 @@ public class KeyedStream<T, KEY> extends DataStream<T> {
 	 *
 	 * @return The transformed {@link DataStream}.
 	 */
+	@Override
 	@PublicEvolving
 	public <R> SingleOutputStreamOperator<R> process(ProcessFunction<T, R> processFunction) {
 
@@ -219,13 +220,14 @@ public class KeyedStream<T, KEY> extends DataStream<T> {
 	 *
 	 * @return The transformed {@link DataStream}.
 	 */
+	@Override
 	@Internal
 	public <R> SingleOutputStreamOperator<R> process(
 			ProcessFunction<T, R> processFunction,
 			TypeInformation<R> outputType) {
 
-		ProcessOperator<KEY, T, R> operator =
-				new ProcessOperator<>(clean(processFunction));
+		KeyedProcessOperator<KEY, T, R> operator =
+				new KeyedProcessOperator<>(clean(processFunction));
 
 		return transform("Process", outputType, operator);
 	}
