+++ flink-streaming-java/src/test/java/org/apache/flink/streaming/api/operators/ProcessOperatorTest.java
@@ -18,16 +18,10 @@
 package org.apache.flink.streaming.api.operators;
 
 
-import org.apache.flink.api.common.state.ValueStateDescriptor;
-import org.apache.flink.api.common.typeinfo.BasicTypeInfo;
-import org.apache.flink.api.common.typeutils.base.IntSerializer;
-import org.apache.flink.api.java.functions.KeySelector;
 import org.apache.flink.streaming.api.TimeDomain;
 import org.apache.flink.streaming.api.functions.ProcessFunction;
 import org.apache.flink.streaming.api.watermark.Watermark;
 import org.apache.flink.streaming.runtime.streamrecord.StreamRecord;
-import org.apache.flink.streaming.runtime.tasks.OperatorStateHandles;
-import org.apache.flink.streaming.util.KeyedOneInputStreamOperatorTestHarness;
 import org.apache.flink.streaming.util.OneInputStreamOperatorTestHarness;
 import org.apache.flink.streaming.util.TestHarnessUtil;
 import org.apache.flink.util.Collector;
@@ -35,8 +29,6 @@ import org.apache.flink.util.TestLogger;
 import org.junit.Test;
 
 import java.util.concurrent.ConcurrentLinkedQueue;
-
-import static org.junit.Assert.assertEquals;
 
 /**
  * Tests {@link ProcessOperator}.
@@ -46,11 +38,11 @@ public class ProcessOperatorTest extends TestLogger {
 	@Test
 	public void testTimestampAndWatermarkQuerying() throws Exception {
 
-		ProcessOperator<Integer, Integer, String> operator =
-				new ProcessOperator<>(new QueryingFlatMapFunction(TimeDomain.EVENT_TIME));
+		ProcessOperator<Integer, String> operator =
+				new ProcessOperator<>(new QueryingProcessFunction(TimeDomain.EVENT_TIME));
 
 		OneInputStreamOperatorTestHarness<Integer, String> testHarness =
-				new KeyedOneInputStreamOperatorTestHarness<>(operator, new IdentityKeySelector<Integer>(), BasicTypeInfo.INT_TYPE_INFO);
+				new OneInputStreamOperatorTestHarness<>(operator);
 
 		testHarness.setup();
 		testHarness.open();
@@ -76,11 +68,11 @@ public class ProcessOperatorTest extends TestLogger {
 	@Test
 	public void testTimestampAndProcessingTimeQuerying() throws Exception {
 
-		ProcessOperator<Integer, Integer, String> operator =
-				new ProcessOperator<>(new QueryingFlatMapFunction(TimeDomain.PROCESSING_TIME));
+		ProcessOperator<Integer, String> operator =
+				new ProcessOperator<>(new QueryingProcessFunction(TimeDomain.PROCESSING_TIME));
 
 		OneInputStreamOperatorTestHarness<Integer, String> testHarness =
-				new KeyedOneInputStreamOperatorTestHarness<>(operator, new IdentityKeySelector<Integer>(), BasicTypeInfo.INT_TYPE_INFO);
+				new OneInputStreamOperatorTestHarness<>(operator);
 
 		testHarness.setup();
 		testHarness.open();
@@ -100,198 +92,13 @@ public class ProcessOperatorTest extends TestLogger {
 
 		testHarness.close();
 	}
-
-	@Test
-	public void testEventTimeTimers() throws Exception {
-
-		ProcessOperator<Integer, Integer, Integer> operator =
-				new ProcessOperator<>(new TriggeringFlatMapFunction(TimeDomain.EVENT_TIME));
-
-		OneInputStreamOperatorTestHarness<Integer, Integer> testHarness =
-				new KeyedOneInputStreamOperatorTestHarness<>(operator, new IdentityKeySelector<Integer>(), BasicTypeInfo.INT_TYPE_INFO);
-
-		testHarness.setup();
-		testHarness.open();
-
-		testHarness.processWatermark(new Watermark(0));
-
-		testHarness.processElement(new StreamRecord<>(17, 42L));
-
-		testHarness.processWatermark(new Watermark(5));
-
-		ConcurrentLinkedQueue<Object> expectedOutput = new ConcurrentLinkedQueue<>();
-
-		expectedOutput.add(new Watermark(0L));
-		expectedOutput.add(new StreamRecord<>(17, 42L));
-		expectedOutput.add(new StreamRecord<>(1777, 5L));
-		expectedOutput.add(new Watermark(5L));
-
-		TestHarnessUtil.assertOutputEquals("Output was not correct.", expectedOutput, testHarness.getOutput());
-
-		testHarness.close();
-	}
-
-	@Test
-	public void testProcessingTimeTimers() throws Exception {
-
-		ProcessOperator<Integer, Integer, Integer> operator =
-				new ProcessOperator<>(new TriggeringFlatMapFunction(TimeDomain.PROCESSING_TIME));
-
-		OneInputStreamOperatorTestHarness<Integer, Integer> testHarness =
-				new KeyedOneInputStreamOperatorTestHarness<>(operator, new IdentityKeySelector<Integer>(), BasicTypeInfo.INT_TYPE_INFO);
-
-		testHarness.setup();
-		testHarness.open();
-
-		testHarness.processElement(new StreamRecord<>(17));
-
-		testHarness.setProcessingTime(5);
-
-		ConcurrentLinkedQueue<Object> expectedOutput = new ConcurrentLinkedQueue<>();
-
-		expectedOutput.add(new StreamRecord<>(17));
-		expectedOutput.add(new StreamRecord<>(1777, 5L));
-
-		TestHarnessUtil.assertOutputEquals("Output was not correct.", expectedOutput, testHarness.getOutput());
-
-		testHarness.close();
-	}
-
-	/**
-	 * Verifies that we don't have leakage between different keys.
-	 */
-	@Test
-	public void testEventTimeTimerWithState() throws Exception {
-
-		ProcessOperator<Integer, Integer, String> operator =
-				new ProcessOperator<>(new TriggeringStatefulFlatMapFunction(TimeDomain.EVENT_TIME));
-
-		OneInputStreamOperatorTestHarness<Integer, String> testHarness =
-				new KeyedOneInputStreamOperatorTestHarness<>(operator, new IdentityKeySelector<Integer>(), BasicTypeInfo.INT_TYPE_INFO);
-
-		testHarness.setup();
-		testHarness.open();
-
-		testHarness.processWatermark(new Watermark(1));
-		testHarness.processElement(new StreamRecord<>(17, 0L)); // should set timer for 6
-
-		testHarness.processWatermark(new Watermark(2));
-		testHarness.processElement(new StreamRecord<>(42, 1L)); // should set timer for 7
-
-		testHarness.processWatermark(new Watermark(6));
-		testHarness.processWatermark(new Watermark(7));
-
-		ConcurrentLinkedQueue<Object> expectedOutput = new ConcurrentLinkedQueue<>();
-
-		expectedOutput.add(new Watermark(1L));
-		expectedOutput.add(new StreamRecord<>("INPUT:17", 0L));
-		expectedOutput.add(new Watermark(2L));
-		expectedOutput.add(new StreamRecord<>("INPUT:42", 1L));
-		expectedOutput.add(new StreamRecord<>("STATE:17", 6L));
-		expectedOutput.add(new Watermark(6L));
-		expectedOutput.add(new StreamRecord<>("STATE:42", 7L));
-		expectedOutput.add(new Watermark(7L));
-
-		TestHarnessUtil.assertOutputEquals("Output was not correct.", expectedOutput, testHarness.getOutput());
-
-		testHarness.close();
-	}
-
-	/**
-	 * Verifies that we don't have leakage between different keys.
-	 */
-	@Test
-	public void testProcessingTimeTimerWithState() throws Exception {
-
-		ProcessOperator<Integer, Integer, String> operator =
-				new ProcessOperator<>(new TriggeringStatefulFlatMapFunction(TimeDomain.PROCESSING_TIME));
-
-		OneInputStreamOperatorTestHarness<Integer, String> testHarness =
-				new KeyedOneInputStreamOperatorTestHarness<>(operator, new IdentityKeySelector<Integer>(), BasicTypeInfo.INT_TYPE_INFO);
-
-		testHarness.setup();
-		testHarness.open();
-
-		testHarness.setProcessingTime(1);
-		testHarness.processElement(new StreamRecord<>(17)); // should set timer for 6
-
-		testHarness.setProcessingTime(2);
-		testHarness.processElement(new StreamRecord<>(42)); // should set timer for 7
-
-		testHarness.setProcessingTime(6);
-		testHarness.setProcessingTime(7);
-
-		ConcurrentLinkedQueue<Object> expectedOutput = new ConcurrentLinkedQueue<>();
-
-		expectedOutput.add(new StreamRecord<>("INPUT:17"));
-		expectedOutput.add(new StreamRecord<>("INPUT:42"));
-		expectedOutput.add(new StreamRecord<>("STATE:17", 6L));
-		expectedOutput.add(new StreamRecord<>("STATE:42", 7L));
-
-		TestHarnessUtil.assertOutputEquals("Output was not correct.", expectedOutput, testHarness.getOutput());
-
-		testHarness.close();
-	}
-
-	@Test
-	public void testSnapshotAndRestore() throws Exception {
-
-		ProcessOperator<Integer, Integer, String> operator =
-				new ProcessOperator<>(new BothTriggeringFlatMapFunction());
-
-		OneInputStreamOperatorTestHarness<Integer, String> testHarness =
-				new KeyedOneInputStreamOperatorTestHarness<>(operator, new IdentityKeySelector<Integer>(), BasicTypeInfo.INT_TYPE_INFO);
-
-		testHarness.setup();
-		testHarness.open();
-
-		testHarness.processElement(new StreamRecord<>(5, 12L));
-
-		// snapshot and restore from scratch
-		OperatorStateHandles snapshot = testHarness.snapshot(0, 0);
-
-		testHarness.close();
-
-		operator = new ProcessOperator<>(new BothTriggeringFlatMapFunction());
-
-		testHarness = new KeyedOneInputStreamOperatorTestHarness<>(operator, new IdentityKeySelector<Integer>(), BasicTypeInfo.INT_TYPE_INFO);
-
-		testHarness.setup();
-		testHarness.initializeState(snapshot);
-		testHarness.open();
-
-		testHarness.setProcessingTime(5);
-		testHarness.processWatermark(new Watermark(6));
-
-		ConcurrentLinkedQueue<Object> expectedOutput = new ConcurrentLinkedQueue<>();
-
-		expectedOutput.add(new StreamRecord<>("PROC:1777", 5L));
-		expectedOutput.add(new StreamRecord<>("EVENT:1777", 6L));
-		expectedOutput.add(new Watermark(6));
-
-		System.out.println("GOT: " + testHarness.getOutput());
-
-		TestHarnessUtil.assertOutputEquals("Output was not correct.", expectedOutput, testHarness.getOutput());
-
-		testHarness.close();
-	}
-
-	private static class IdentityKeySelector<T> implements KeySelector<T, T> {
-		private static final long serialVersionUID = 1L;
-
-		@Override
-		public T getKey(T value) throws Exception {
-			return value;
-		}
-	}
-
-	private static class QueryingFlatMapFunction extends ProcessFunction<Integer, String> {
+	private static class QueryingProcessFunction extends ProcessFunction<Integer, String> {
 
 		private static final long serialVersionUID = 1L;
 
 		private final TimeDomain timeDomain;
 
-		public QueryingFlatMapFunction(TimeDomain timeDomain) {
+		public QueryingProcessFunction(TimeDomain timeDomain) {
 			this.timeDomain = timeDomain;
 		}
 
@@ -311,93 +118,4 @@ public class ProcessOperatorTest extends TestLogger {
 				Collector<String> out) throws Exception {
 		}
 	}
-
-	private static class TriggeringFlatMapFunction extends ProcessFunction<Integer, Integer> {
-
-		private static final long serialVersionUID = 1L;
-
-		private final TimeDomain timeDomain;
-
-		public TriggeringFlatMapFunction(TimeDomain timeDomain) {
-			this.timeDomain = timeDomain;
-		}
-
-		@Override
-		public void processElement(Integer value, Context ctx, Collector<Integer> out) throws Exception {
-			out.collect(value);
-			if (timeDomain.equals(TimeDomain.EVENT_TIME)) {
-				ctx.timerService().registerEventTimeTimer(ctx.timerService().currentWatermark() + 5);
-			} else {
-				ctx.timerService().registerProcessingTimeTimer(ctx.timerService().currentProcessingTime() + 5);
-			}
-		}
-
-		@Override
-		public void onTimer(
-				long timestamp,
-				OnTimerContext ctx,
-				Collector<Integer> out) throws Exception {
-
-			assertEquals(this.timeDomain, ctx.timeDomain());
-			out.collect(1777);
-		}
-	}
-
-	private static class TriggeringStatefulFlatMapFunction extends ProcessFunction<Integer, String> {
-
-		private static final long serialVersionUID = 1L;
-
-		private final ValueStateDescriptor<Integer> state =
-				new ValueStateDescriptor<>("seen-element", IntSerializer.INSTANCE);
-
-		private final TimeDomain timeDomain;
-
-		public TriggeringStatefulFlatMapFunction(TimeDomain timeDomain) {
-			this.timeDomain = timeDomain;
-		}
-
-		@Override
-		public void processElement(Integer value, Context ctx, Collector<String> out) throws Exception {
-			out.collect("INPUT:" + value);
-			getRuntimeContext().getState(state).update(value);
-			if (timeDomain.equals(TimeDomain.EVENT_TIME)) {
-				ctx.timerService().registerEventTimeTimer(ctx.timerService().currentWatermark() + 5);
-			} else {
-				ctx.timerService().registerProcessingTimeTimer(ctx.timerService().currentProcessingTime() + 5);
-			}
-		}
-
-		@Override
-		public void onTimer(
-				long timestamp,
-				OnTimerContext ctx,
-				Collector<String> out) throws Exception {
-			assertEquals(this.timeDomain, ctx.timeDomain());
-			out.collect("STATE:" + getRuntimeContext().getState(state).value());
-		}
-	}
-
-	private static class BothTriggeringFlatMapFunction extends ProcessFunction<Integer, String> {
-
-		private static final long serialVersionUID = 1L;
-
-		@Override
-		public void processElement(Integer value, Context ctx, Collector<String> out) throws Exception {
-			ctx.timerService().registerProcessingTimeTimer(5);
-			ctx.timerService().registerEventTimeTimer(6);
-		}
-
-		@Override
-		public void onTimer(
-				long timestamp,
-				OnTimerContext ctx,
-				Collector<String> out) throws Exception {
-			if (TimeDomain.EVENT_TIME.equals(ctx.timeDomain())) {
-				out.collect("EVENT:1777");
-			} else {
-				out.collect("PROC:1777");
-			}
-		}
-	}
-
 }
